# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrocarLuz API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Lead {
  id: string;
  type: string;
  nome: string;
  whatsapp: string;
  estado: string;
  /** @nullable */
  distribuidora?: string | null;
  /** @nullable */
  consumoBand?: string | null;
  /** @nullable */
  empresa?: string | null;
  /** @nullable */
  valorContaBand?: string | null;
  /** @nullable */
  segmento?: string | null;
  /** @nullable */
  pageSource?: string | null;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  status: string;
  createdAt: string;
}

export type LeadInputType = typeof LeadInputType[keyof typeof LeadInputType];


export const LeadInputType = {
  residential: 'residential',
  business: 'business',
} as const;

export interface LeadInput {
  type: LeadInputType;
  /** @minLength 2 */
  nome: string;
  /** @minLength 8 */
  whatsapp: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  estado: string;
  distribuidora?: string;
  consumoBand?: string;
  empresa?: string;
  valorContaBand?: string;
  segmento?: string;
  pageSource?: string;
  utmSource?: string;
  utmMedium?: string;
  utmCampaign?: string;
}

export interface ArticleSummary {
  id: string;
  slug: string;
  title: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  estado?: string | null;
  author: string;
  /** @nullable */
  publishedAt?: string | null;
  updatedAt: string;
}

/**
 * @nullable
 */
export type ArticleFaqItems = { [key: string]: unknown } | null;

export interface Article {
  id: string;
  slug: string;
  title: string;
  /** @nullable */
  metaDescription?: string | null;
  content: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  estado?: string | null;
  keywords?: string[];
  author: string;
  /** @nullable */
  reviewer?: string | null;
  /** @nullable */
  schemaType?: string | null;
  /** @nullable */
  faqItems?: ArticleFaqItems;
  published: boolean;
  /** @nullable */
  publishedAt?: string | null;
  updatedAt: string;
  createdAt: string;
}

export type ArticleInputFaqItems = { [key: string]: unknown };

export interface ArticleInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  title: string;
  metaDescription?: string;
  /** @minLength 1 */
  content: string;
  excerpt?: string;
  category?: string;
  estado?: string;
  keywords?: string[];
  author?: string;
  reviewer?: string;
  schemaType?: string;
  faqItems?: ArticleInputFaqItems;
  published?: boolean;
}

export interface Faq {
  id: string;
  question: string;
  answer: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  displayOrder?: number | null;
  published: boolean;
  createdAt: string;
}

export interface FaqInput {
  /** @minLength 1 */
  question: string;
  /** @minLength 1 */
  answer: string;
  category?: string;
  displayOrder?: number;
  published?: boolean;
}

export interface State {
  slug: string;
  name: string;
  uf: string;
  region: string;
  distributor: string;
  /** @nullable */
  avgTariffResidential?: number | null;
  gdAvailable: boolean;
  aclAvailable2028: boolean;
  /** @nullable */
  description?: string | null;
}

export type ListArticlesParams = {
category?: string;
estado?: string;
limit?: number;
offset?: number;
};

export type ListFaqsParams = {
category?: string;
};