# lib/api-zod/src/generated/types/article.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrocarLuz API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleFaqItems } from './articleFaqItems';

export interface Article {
  id: string;
  slug: string;
  title: string;
  /** @nullable */
  metaDescription?: string | null;
  content: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  estado?: string | null;
  keywords?: string[];
  author: string;
  /** @nullable */
  reviewer?: string | null;
  /** @nullable */
  schemaType?: string | null;
  /** @nullable */
  faqItems?: ArticleFaqItems;
  published: boolean;
  /** @nullable */
  publishedAt?: string | null;
  updatedAt: string;
  createdAt: string;
}