# lib/api-zod/src/generated/types/articleInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrocarLuz API
 * OpenAPI spec version: 0.1.0
 */
import type { ArticleInputFaqItems } from './articleInputFaqItems';

export interface ArticleInput {
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  title: string;
  metaDescription?: string;
  /** @minLength 1 */
  content: string;
  excerpt?: string;
  category?: string;
  estado?: string;
  keywords?: string[];
  author?: string;
  reviewer?: string;
  schemaType?: string;
  faqItems?: ArticleInputFaqItems;
  published?: boolean;
}