# lib/api-zod/src/generated/types/articleSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrocarLuz API
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleSummary {
  id: string;
  slug: string;
  title: string;
  /** @nullable */
  excerpt?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  estado?: string | null;
  author: string;
  /** @nullable */
  publishedAt?: string | null;
  updatedAt: string;
}