# lib/api-zod/src/generated/types/faq.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrocarLuz API
 * OpenAPI spec version: 0.1.0
 */

export interface Faq {
  id: string;
  question: string;
  answer: string;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  displayOrder?: number | null;
  published: boolean;
  createdAt: string;
}