# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrocarLuz API
 * OpenAPI spec version: 0.1.0
 */

export interface Lead {
  id: string;
  type: string;
  nome: string;
  whatsapp: string;
  estado: string;
  /** @nullable */
  distribuidora?: string | null;
  /** @nullable */
  consumoBand?: string | null;
  /** @nullable */
  empresa?: string | null;
  /** @nullable */
  valorContaBand?: string | null;
  /** @nullable */
  segmento?: string | null;
  /** @nullable */
  pageSource?: string | null;
  /** @nullable */
  utmSource?: string | null;
  /** @nullable */
  utmMedium?: string | null;
  /** @nullable */
  utmCampaign?: string | null;
  status: string;
  createdAt: string;
}