# lib/api-zod/src/generated/types/leadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrocarLuz API
 * OpenAPI spec version: 0.1.0
 */
import type { LeadInputType } from './leadInputType';

export interface LeadInput {
  type: LeadInputType;
  /** @minLength 2 */
  nome: string;
  /** @minLength 8 */
  whatsapp: string;
  /**
     * @minLength 2
     * @maxLength 2
     */
  estado: string;
  distribuidora?: string;
  consumoBand?: string;
  empresa?: string;
  valorContaBand?: string;
  segmento?: string;
  pageSource?: string;
  utmSource?: string;
  utmMedium?: string;
  utmCampaign?: string;
}