# lib/api-zod/src/generated/types/state.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrocarLuz API
 * OpenAPI spec version: 0.1.0
 */

export interface State {
  slug: string;
  name: string;
  uf: string;
  region: string;
  distributor: string;
  /** @nullable */
  avgTariffResidential?: number | null;
  gdAvailable: boolean;
  aclAvailable2028: boolean;
  /** @nullable */
  description?: string | null;
}